Added a _groups filter. Check one or more groups and click 'filter,' and you'll see only users who are a member of at least one of those groups. Clear the group checkboxes to view all users again. I've kept the changes minimal (just one new file, and an overridden action). The trickiest bit is preventing the base class from executing invalid filter code for groups.
I added _groups to the default list of filters in generator.yml, because
it's darn useful, but we could just document it in the README instead.

I've taken steps to prevent this from breaking in future versions of
PHP that might be pickier about compiling references to undefined constants
(even if that code never actually runs).

I haven't tried to merge this into the 1.1 and 1.2 branches, they
seem to be similar, it should be easy to do.

-TBB (tom@punkave.com)



git-svn-id: http://svn.symfony-project.com/plugins/sfGuardPlugin/branches/1.0@10861 ee427ae8-e902-0410-961c-c3ed070cd9f9